new: add new custom surfaces
