+++ pygaming/__init__.py
@@ -17,8 +17,9 @@ from .transition import ServerTransition, GameTransition
 
 from .screen.screen import Screen
 from .screen.frame import Frame
-from .screen.element import Element, TOP_LEFT, TOP_RIGHT, CENTER, BOTTOM_LEFT, BOTTOM_RIGHT
-from .screen.animated_surface import AnimatedSurface, SurfaceLike, SurfacesLike
+from .screen.element import Element, TOP_LEFT, TOP_RIGHT, CENTER, BOTTOM_LEFT, BOTTOM_RIGHT, SurfaceLike
+from .screen.animated_surface import AnimatedSurface, SurfacesLike
+from .screen.label import Label, TEXT_CENTERED, TEXT_LEFT, TEXT_RIGHT
 
 from .inputs import Inputs, Controls, Click, Keyboard, Mouse
 from .file import FontFile, DataFile, ImageFile, GIFFile, SoundFile, MusicFile, get_file
@@ -26,11 +27,12 @@ from .connexion import Client, Server as Network, HEADER, ID, CONTENT, TIMESTAMP
 
 from .database import Database, Texts, Speeches
 from . import commands
-from .screen.colored_surface import ColoredSurface
+from .screen.colored_surfaces import ColoredRectangle
 
 __all__ = ['Config', 'Font', 'Game', 'NO_NEXT', 'STAY', 'Logger', 'ServerPhase', 'GamePhase',
            'Server', 'Settings', 'ServerTransition', 'GameTransition', 'Screen', 'Frame',
            'Element', 'AnimatedSurface', 'SurfaceLike', 'SurfacesLike', 'Inputs', 'Controls', 'Click', 'FontFile',
            'DataFile', 'ImageFile', 'GIFFile', 'SoundFile', 'MusicFile','get_file', 'Client', 'Keyboard', 'Mouse',
            'Network', 'HEADER', 'ID', 'CONTENT', 'TIMESTAMP', 'Database', 'Texts', 'Speeches',
-           'commands', 'ColoredSurface', 'TOP_LEFT', 'TOP_RIGHT', 'CENTER', 'BOTTOM_LEFT', 'BOTTOM_RIGHT']
+           'commands', 'ColoredRectangle', 'TOP_LEFT', 'TOP_RIGHT', 'CENTER', 'BOTTOM_LEFT', 'BOTTOM_RIGHT',
+           'Label', 'TEXT_CENTERED', 'TEXT_LEFT', 'TEXT_RIGHT']
